+++ src/index.ts
@@ -3,6 +3,12 @@ export * from './parser'
 export * from './models'
 export * from './errors'
 
+// Utility functions
+export * from './utils'
+
 // Optional typing layer
 export * from './converters'
 export * from './validators'
+
+// Type definitions
+export * from './types'